divide test (#678)
