+++ jest.config.cjs
@@ -154,15 +154,15 @@ module.exports = {
 
 	// The glob patterns Jest uses to detect test files
 	testMatch: [
-		"**/__tests__/**/*.[jt]s?(x)",
-		"**/?(*.)+(spec|test).[tj]s?(x)",
+		// "**/__tests__/**/*.[jt]s?(x)",
+		// "**/?(*.)+(spec|test).[tj]s?(x)",
 		"<rootDir>/test/**/*"
 	],
 
 	// An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-	// testPathIgnorePatterns: [
-	//   "\\\\node_modules\\\\"
-	// ],
+	testPathIgnorePatterns: [
+		"<rootDir>/test/testutils.ts",
+	],
 
 	// The regexp pattern or array of patterns that Jest uses to detect test files
 	// testRegex: [],
